requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
scikit-learn
streamlit
requests
python-multipart
python-docx
